requirements.txt
pydantic
streamlit
python-dotenv
llama-index-core
llama-index-llms-openai
llama-index-readers-file